% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data.R
\name{calculate_rasta}
\alias{calculate_rasta}
\title{Calculate a specific RASTA(A, B, C, Ref) from an input block table.}
\usage{
calculate_rasta(raw_ras, A, B, C)
}
\arguments{
\item{raw_ras}{tibble. A tibble containing the results}

\item{A}{character. Population A fr the rasta calculation. Must be present in the raw_ras$Left.}

\item{B}{character. Population B for the rasta calculation. Must be present in the raw_ras$Left.}

\item{C}{character. Population C for the rasta calculation. Must be present in the raw_ras$Right.}
}
\value{
A tibble with the requested RASTA, it's error and ZScore
}
\description{
Calculate a specific RASTA(A, B, C, Ref) from an input block table.
}
