% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data.R
\name{check_x_in_y}
\alias{check_x_in_y}
\title{Check that x is not empty and all elements exist in y.}
\usage{
check_x_in_y(x, y, var_name)
}
\arguments{
\item{x}{character. A vector of strings that must exist in y.}

\item{y}{character. A vector of character that must include all elements in x.}

\item{var_name}{character. The variable name to be included in error messages.}
}
\description{
Check that x is not empty and all elements exist in y.
}
