% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{delete_mj_jackknife}
\alias{delete_mj_jackknife}
\title{Delete mj jackknife function}
\usage{
delete_mj_jackknife(data, values, mj, weights, .print_vals = F)
}
\arguments{
\item{data}{Tibble containing the data to use.}

\item{values}{Unquoted name of the tibble column with the block values.}

\item{mj}{Unquoted name of the tibble column with the mj values.}

\item{weights}{Unquoted name of the tibble column with the block weights.}

\item{.print_vals}{logical. Should the actual theta_minus values be printed?}
}
\value{
list. Returns a list of items :
theta_J  The jackknife mean from all the theta_-x data.
jack_se The SE of theta_J
Zscore  The Zscore of the theta_J
jack.values A vector of all the theta_-x values
}
\description{
Implementation of delete mj jackknifing following Nick Patterson's writeup of it.
The function works like tidyverse functions, i.e. values, mj and weights should not be quoted/be symbols.
}
