% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats_toolkit.R
\name{f3_to_similarity_matrix}
\alias{f3_to_similarity_matrix}
\title{f3_to_similarity_matrix}
\usage{
f3_to_similarity_matrix(f3_table, na_diag = F)
}
\arguments{
\item{f3_table}{tibble. A tibble containing the contents of the F3tableOutFile.}

\item{na_diag}{logical. If TRUE the diagonal of the similarity matrix is set to NA.}
}
\value{
tibble. A similarity matrix where the 'b' populations are columns and their values are the F3 of each 'a' with each 'b'. Standard errors are discarded.
}
\description{
f3_to_similarity_matrix
}
