% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{k_nearest_neighbour}
\alias{k_nearest_neighbour}
\title{K Nearest Neighbour Classification}
\usage{
k_nearest_neighbour(x, group_definitions, k = 5, exclude_self_sharing = T)
}
\arguments{
\item{x}{matrix. The input similarity matrix.}

\item{group_definitions}{tibble. A tibble with two columns that specifies the name of each column in the input matrix and the grouping it belongs to.}

\item{k}{integer. The number of nearest neighbours to use for classification.}

\item{exclude_self_sharing}{logical. Should self sharing be excluded? If so, the highest sharing in each row is ignored (assumed to be self-sharing).}
}
\value{
Tibble. A tibble containing the assigned group of the k nearest neighbours as well as the accuracy of the knn at assigning the correct
grouping based on the similarity scores for each row in the input matrix.
}
\description{
Preform k-nearest-neighbour classification from a similarity matrix and a list classifying the columns into groups.
}
