% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data.R
\name{ras_to_rasta}
\alias{ras_to_rasta}
\title{A function to calculate RASTA(A, B; C, Ref) from a blockTable of RAS per chromosome.}
\usage{
ras_to_rasta(block_table, A, B, C)
}
\arguments{
\item{block_table}{tibble. A tibble with the contents of a blockTableFile, derived with read_xerxes_ras_blocks()}

\item{A}{character. The Lefts to cycle as population A.}

\item{B}{character. The Lefts to cycle as population B.}

\item{C}{character. The Rights to cycle as population C.}
}
\value{
A tibble with the all the requested combinations of rasta, along with their SE and Zscores
}
\description{
A function to calculate RASTA(A, B; C, Ref) from a blockTable of RAS per chromosome.
}
