% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ras_toolkit.R
\name{ras_to_similarity_matrix}
\alias{ras_to_similarity_matrix}
\title{ras_to_similarity_matrix}
\usage{
ras_to_similarity_matrix(ras_table, na_diag = F)
}
\arguments{
\item{ras_table}{tibble. A tibble containing the contents of the tableOutFile.}

\item{na_diag}{logical. If TRUE the diagonal of the similarity matrix is set to NA.}
}
\value{
tibble. A similarity matrix where the Right populations are columns and their values are the ras of each Left with each right. Standard errors are discarded.
}
\description{
ras_to_similarity_matrix
}
