% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read_xerxes_f3_table}
\alias{read_xerxes_f3_table}
\title{read_xerxes_f3_table}
\usage{
read_xerxes_f3_table(fn, strip_selection_language = F)
}
\arguments{
\item{fn}{path. The path to the input tableOutFile.}

\item{strip_selection_language}{logical. Should trident selection language marks be stripped from columns a,b and c?
If true, selection marks such as "<" and ">" will be stripped from the a, b and c columns.}
}
\value{
A tibble containing the contents of the tableOutFile.
}
\description{
read_xerxes_f3_table
}
