% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read_xerxes_rasf4_table}
\alias{read_xerxes_rasf4_table}
\title{read_xerxes_rasf4_table}
\usage{
read_xerxes_rasf4_table(fn, strip_selection_language = F)
}
\arguments{
\item{fn}{character. The path to the input f4TableOutFile.}

\item{strip_selection_language}{logical. Should trident selection language marks be stripped from Lefts and Rights?
If true, selection marks such as "<" and ">" will be stripped from the Left and Right columns.}
}
\value{
A tibble containing the contents of the f4TableOutFile.
}
\description{
Read xerxes ras f4TableOutFile contents into a tibble.
}
